requirements.txt
amqp==2.5.2
aniso8601==8.0.0
appdirs==1.4.4
asgiref==3.2.3
autopep8==1.5
Babel==2.8.0
beautifulsoup4==4.8.2
billiard==3.6.1.0
bleach==3.1.0
boto3==1.11.6
botocore==1.14.6
braintree==3.58.0
cachetools==4.0.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cssselect2==0.2.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
dj-email-url==0.2.0
Django==3.0.2
django-appconf==1.0.3
django-archive==0.1.6
django-babel==0.6.2
django-bootstrap-modal-forms==1.5.0
django-compressor==2.4
django-cors-headers==3.2.1
django-countries==5.5
django-db-logger==0.1.7
django-filter==2.2.0
django-graphql-jwt==0.3.0
django-js-asset==1.2.2
django-measurement==3.2.0
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-prettyjson==0.4.1
django-prices==2.1.0
django-prices-openexchangerates==1.0.1
django-prices-vatlayer==1.0.2
django-recaptcha==2.0.6
django-render-block==0.6
django-robots==4.0
django-silk==4.0.0
django-storages==1.8
django-tables2==2.2.1
django-templated-email==2.3.0
django-tinymce4==3.0.2
django-tinymce4-lite==1.7.5
django-user-agents==0.4.0
django-versatileimagefield==2.0
djangorestframework==3.11.0
docutils==0.16
draftjs-sanitizer==1.0.0
Faker==4.0.0
filelock==3.0.12
freezegun==0.3.13
geographiclib==1.50
geopy==1.22.0
google-api-core==1.16.0
google-auth==1.10.1
google-cloud-core==1.2.0
google-cloud-storage==1.25.0
google-i18n-address==2.3.5
google-measurement-protocol==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gprof2dot==2019.11.30
graphene==2.1.8
graphene-django==2.8.0
graphene-django-optimizer==0.6.1
graphene-federation==0.0.6
graphql-core==3.0.1
graphql-relay==3.0.0
html-to-draftjs==1.0.1
html5lib==1.0.1
idna==2.8
importlib-metadata==3.4.0
Jinja2==2.10.3
jmespath==0.9.4
jsmin==2.2.2
jsonfield==2.0.2
kombu==4.6.7
libmagic==1.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
maxminddb==1.5.2
maxminddb-geolite2==2018.703
measurement==3.2.0
more-itertools==8.1.0
mpmath==1.1.0
mysqlclient==1.4.6
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
pandas==0.25.3
phonenumberslite==8.11.2
Pillow==7.0.0
prices==1.0.0
promise==2.3
protobuf==3.11.2
psycopg2-binary==2.8.4
purl==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.5.2
PyJWT==1.7.1
Pyphen==0.9.5
python-dateutil==2.8.1
python-magic==0.4.15
python-memcached==1.59
pytz==2019.3
razorpay==1.2.0
rcssmin==1.0.6
redis==3.3.11
requests==2.22.0
rjsmin==1.1.0
rsa==4.0
Rx==3.0.1
s3transfer==0.3.1
sentry-sdk==0.14.0
singledispatch==3.4.0.3
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
standardjson==0.3.1
stripe==2.42.0
sympy==1.5.1
text-unidecode==1.3
tinycss2==1.0.2
tqdm==4.41.1
typing==3.7.4.1
typing-extensions==3.7.4.3
ua-parser==0.8.0
urllib3==1.25.7
user-agents==2.0
uWSGI==2.0.18
vine==1.3.0
virtualenv==20.4.2
WeasyPrint==51
webencodings==0.5.1
zipp==3.4.0
